requirements.txt
aiogram==2.25.1
Pyrogram==2.0.106

flake8==6.1.0
isort==5.12.0

psycopg2-binary==2.9.9
TgCrypto==1.2.5
python-dotenv==1.0.0

